+++ apps/bot/src/lib/agent/types/errors.ts
@@ -1,0 +1,1 @@
+export type AgentErrorStage = 'after_progress' | 'after_text' | 'before_output';
