+++ apps/bot/src/lib/ai/attempts.ts
@@ -1,9 +1,5 @@
 import type { PiAttempt } from '@repo/ai';
-
-export interface AttemptFailure {
-  attempt: PiAttempt;
-  error: unknown;
-}
+import type { AttemptFailure } from '@/lib/ai/types/attempts';
 
 export function nextAttempt({
   attempts,
