+++ apps/bot/src/lib/ai/stream/tasks/default.ts
@@ -1,5 +1,5 @@
 import { errorOutput, textField } from './helpers';
-import type { DefaultToolTaskRenderer } from './types';
+import type { DefaultToolTaskRenderer } from './types/renderers';
 
 export const defaultTool: DefaultToolTaskRenderer = {
   request: ({ input, toolName }) => {
