+++ apps/bot/src/lib/ai/stream/tasks/summarize-thread.ts
@@ -1,5 +1,5 @@
 import { numberField, plural, textField } from './helpers';
-import type { ToolTaskRendererEntry } from './types';
+import type { ToolTaskRendererEntry } from './types/renderers';
 
 export const summarizeThread: ToolTaskRendererEntry = {
   title: 'Summarizing thread',
