+++ apps/bot/src/lib/ai/stream/tasks/types/renderers.ts
