+++ apps/bot/src/lib/ai/types/attachments.ts
@@ -1,0 +1,6 @@
+export interface SeededAttachment {
+  mimeType?: string;
+  name: string;
+  path: string;
+  type: string;
+}
