+++ apps/bot/src/lib/ai/types/attempts.ts
@@ -1,0 +1,6 @@
+import type { PiAttempt } from '@repo/ai';
+
+export interface AttemptFailure {
+  attempt: PiAttempt;
+  error: unknown;
+}
