+++ apps/bot/src/lib/ai/types/tools/generate-image.ts
@@ -1,0 +1,6 @@
+export interface GeneratedImage {
+  bytes: Uint8Array;
+  index: number;
+  mediaType: string;
+  total: number;
+}
