+++ apps/bot/src/lib/errors.ts
@@ -1,3 +1,4 @@
+import type { AgentErrorStage } from '@/lib/agent/types/errors';
 import { errorMessage } from '@/lib/utils/error';
 
 const CREDIT_ERROR_PATTERN =
@@ -6,8 +7,6 @@ const CONTEXT_ERROR_PATTERN =
   /\b(max_tokens|maximum context|context length|too many tokens)\b/i;
 const PROVIDER_TIMEOUT_PATTERN =
   /\b(5\d\d|gateway time-out|gateway timeout|cloudflare|timeout occurred|origin_response_timeout)\b/i;
-
-export type AgentErrorStage = 'after_progress' | 'after_text' | 'before_output';
 
 export function agentErrorMessage({
   error,
