+++ apps/bot/src/lib/slack/types/names.ts
@@ -1,0 +1,13 @@
+interface ProfileField {
+  label: string;
+  value: string;
+}
+
+export interface UserProfile {
+  displayName?: string;
+  fields?: ProfileField[];
+  pronouns?: string;
+  realName?: string;
+  status?: string;
+  title?: string;
+}
